Processes are removed from new list after a certain time, pipe-delimited instead of tab.
To prevent processes which die before they reach the GPU from holding resources forever, there's a new parameter that determines the maximum amount of time that a process will be kept in the new list before it's removed and its resources are made available for others (by default this is two minutes).

There may also have been an error previously where processes would stay in the new list because in nvidia-smi it didn't show " &" at the end (from backgrounding the process), so we didn't recognize that the process had started. Processes are still run in the background, but now we don't add " &" to the command that's stored in the process list.