lock_dir + lock_suffix params for locking functions, functions to return GPU named tuples.
To make the locking functions friendlier for use in other scripts, they now take lock_dir and lock_suffix as inputs. The way signal calls the cleanup function was updated to allow passing these in (we also now cleanup after catching the signal SIGINT).

Because getting GPU information is useful in other places as well, some of that code has been pulled into functions. In particular, there's now nvidia_smi to run the command of that name for you and return the output as a string. There's also get_gpus which returns a named tuple for each GPU with information about the memory and utilization free on that GPU.