Flask and Vue driven web page to start new jobs, view running jobs, and view GPU info.
The web page works fairly well at this point. New jobs can be created (either pasting into the text field or using a nice method of generating all combinations of the parameters that you want to change between each script), and they're added to the jobs file in a way that respects the locks on it (though this hasn't really been tested).

Every 30 seconds (not easily adjustable at present) the page updates a table that shows all of the jobs currently in the file (those waiting to be run on a GPU). There's a nice search bar to filter these if you need to check for certain jobs.

The page has an area where it expects to receive information about each GPU. However, there's just some dummy data right now.

There is no password needed yet, so this isn't very secure; anybody who can access the web page can write arbitrary code which the GPU runner will execute.