write_to_locked_file function.
The code for writing to the jobs file from the server is now in a function in gpu_runner.py (since all of the other functions for working with locked things are there as well). gpu_runner_frontend.py uses this function now which should make it much easier to write to the jobs file in general.