Fixes to skip GPUs properly.
Changes to use dictionaries with GPU number instead of lists. Skipping ought to work now, but still isn't really tested.
Also a check that the number of GPUs matched by memory is the same as those matched by utilization (there might be errors on machines with older GPUs where not all information is available).