Secret key and password creation for new user; fixes #14.
If the secret key or password files don't exist, a secret key will be created or the user will be prompted to enter a username and password. These files are chmod'ed to 600.
Also, if an incorrect password is given, an empty URL is returned instead of trying to re-render the login page. On the frontend, no navigation is done unless the URL is truthy.