Support for GPU usage on multiple machines.
A new add_machine page where you can add necessary information to ssh into a new machine (username, address, path to a password file (to be used with sshpass), and a display name). The machines are displayed in a data table and support searching and deleting machines (editing isn't supported; you have to delete and re-create a machine).

data/gpu runs nvidia-smi on each of the machines. Running it on the current machine (where the server itself is running), is _not_ supported now (because it expects to always SSH somewhere first). This shouldn't be too hard to add if needed; one could create a machine with address localhost and then have the backend run no ssh_command on that machine.

dashboard now shows the GPU data from each machine. However, this is currently just a long list (one can't even tell which GPUs are on which machine). This should be updated soon.