Switched from password file to interactive pw entry.
You have to enter your SSH password when starting the server, but after that it won't be needed again for each use. Note that sshpass with this method could show the password in the SSH command (visible, e.g., when running `ps`) on older systems. On current Ubuntu, the password is censored (to zzzzzzzzzzzzzzz or such).