Persistent SSH connections, able to delete jobs; fixes #7 and fixes #8.
On start, one SSH connection is made to each machine. These are kept the entire time so that we don't have to SSH in again each time that we want to run another command. So far, these are used for getting the GPU data refresh every 30 seconds. ssh.py also defines a second class (SSHBackgroundConnection) that will be used for running jobs on the remote machines; these use a daemon thread to continually process the output, and they can find output lines with a particular tag. This will be used so the models can denote which output lines should be put into the log collection for them in Mongo.

On the dashboard, the jobs are now deletable.