Added GPU runner toggle (frontend only) and input even for switching machines.
There's a toggle to switch the GPU runner on/off for the currently selected machine; this doesn't do anything yet, though, as it's not connected to the backend.

There's also a function that's called now whenever you switch machines. This only logs the name of the new machine to the console, but this will be updated later to give an immediate refresh of jobs data.