Machine class, logging fixes.
ssh.py only defines one class now; there's no longer a second class that tries to handle the logging as well. We've moved away from using stdout to do any logging, so the dedicated class that processed stdout and filtered for the right tags was removed.

gpu_runner_web_interface defines a Machine class. Each Machine owns its own client for running code on that machine. It uses a lock to keep access to this threadsafe so that we only need one connection to each machine (which is used for getting GPU data, starting jobs, and processing logs). Log processing was no longer dependent on the SSH connection after switching to log files which are saved on the remote machine, so there's a function in the web file that takes care of this. A thread is spawned for each Machine which continually runs handle_machine, a new function that does both starting jobs and processing logs. Starting jobs for each machine can be turned on or off (but this isn't integrated with the frontend yet so it's hard-coded that all of them are on).