Further but incomplete GPU runner toggling; also fixes #17.
This hasn't been tested yet, but GPU runner state (on/off) should be saved to the particular machine on the frontend when toggled. This state is passed from the backend when queried for machines. However, there's still no notification from the frontend to the backend of a change in status.

Switching machines should now update the jobs immediately; this has also not been tested yet.