One card per machine for GPU usage.
This makes it so you can finally tell which GPUs are on which machine. Flask has also been configured to check for updated templates, so changing a template after the server has started running will now cause the updated templated to be loaded the next time a render request for it is received (this is much nicer for development and debugging than having to restart the server each time; debug mode should also give this, but that also seems to have the backend restart and require password re-entry. Perhaps that's not the case if only a template is altered?).