Fixed reading logs, gpu (full output), and tmux_gpu_info.
awk seems to use \r\n instead of \n; replacing these fixes reading in the log files.
gpu was using nvidia_smi without full output and thus not showing process information; this works as before now.
tmux_gpu_info and gpu were both updated to use better imports (from gpu_dashboard package).
The repo will be renamed to gpu_dashboard so that the repo name doesn't conflict with the name of one of the files inside of it, which messes up imports if that file is on your path. This also reflects the changing focus of the repository to create more of a web hub for GPU running instead of just a few utility scripts.