Bar charts for GPU usage, formatting error snackbar; fixes #2.
Instead of a purely textual display, there are now two bar charts per GPU on each machine which show the % memory and % utilization being used. Each machine has its own expansion panel which shows these charts in the header (so they're always visible). If the panel is expanded, it just shows the text information that was previously shown for each machine (which shows memory usage in absolute units instead of as a percentage). In the future, this will probably be changed so that expanding shows which processes are running on the GPUs and whose they are.

If the format of the submitted commands isn't mem|util|cmd then a snackbar pops up to alert the user to their error, and the commands aren't submitted.