Ignore machines that can't be connected to.
If there is an error while trying to connect to a given machine, the server will ignore that machine, connect to the others, and run with whatever machines work. This is nice, but it should be improved so that 1) Add Machine shows all machines, even those that one doesn't have a connection to (so that the user doesn't think they need to add the machine again to get it to work) and 2) there should be a way (either after a timeout or when the user specifies it) to try to connect again to the machines that initially didn't work.